refactor: add yarn workspace packagesbase64, three
