+++ package.json
@@ -1,6 +1,13 @@
 {
-  "name": "PanoToMesh",
-  "version": "0.1.0",
+  "name": "pano-to-mesh",
+  "version": "1.2.0",
+  "private": true,
+  "license": "MIT",
+  "workspaces": {
+    "packages": [
+      "packages/*"
+    ]
+  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -8,14 +15,13 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "js-base64": "^3.7.5",
+    "@pano-to-mesh/three": "1.0.0",
+    "@pano-to-mesh/base64": "1.0.0",
     "next": "13.2.4",
-    "pako": "^2.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-rnd": "^10.4.1",
     "styled-components": "^6.0.4",
-    "three": "^0.149.0",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
