+++ packages/three/components/MeshIndexMap/index.js
