+++ packages/three/components/Placeholder/index.js
