+++ packages/three/components/ThreeCanvas/index.js
