+++ packages/three/core/CapturePanorama.js
