+++ packages/three/core/TexturePostEffect.js
