+++ packages/three/hooks/texture.js
