+++ packages/three/package.json
@@ -1,0 +1,9 @@
+{
+  "version": "1.0.0",
+  "name": "@pano-to-mesh/three",
+  "dependencies": {
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "three": "^0.149.0"
+  }
+}
