+++ packages/three/shaders/fragmentShaders/cubemapToEquirectangular.js
