+++ packages/three/shaders/fragmentShaders/edgeDetection.js
