+++ packages/three/shaders/fragmentShaders/equirectangularProjection.js
