+++ packages/three/shaders/fragmentShaders/hoverEffect.js
