+++ packages/three/shaders/fragmentShaders/texture.js
