+++ packages/three/shaders/fragmentShaders/vertexColor.js
