+++ packages/three/shaders/vertexShaders/worldPosition.js
